refactor: move functions from pre-start hook to pouchd
move the functions: `set rootfs quota` and `set mtab` from pre-start
hook in runc to pouchd container's create. Pouchd mounts the rootfs and
do these functions, if it is successful, and then umount the rootfs.
